[refactor] Bean 생성시점 변경
