+++ src/main/java/core/mvc/DispatcherServlet.java
@@ -56,11 +56,10 @@ public class DispatcherServlet extends HttpServlet {
                 return;
             }
 
-
             ModelAndView mav = handlerExecutor.handle(req, resp, maybeHandler.get());
             render(mav, req, resp);
         } catch (Throwable e) {
-            logger.error("Exception : {}", e);
+            logger.error("Exception : ", e);
             throw new ServletException(e.getMessage());
         }
     }
